InMigration report calculation is ready for review.
git-svn-id: http://svn.rcg.usm.maine.edu/svn/openhds/openhds-core@1933 0411eb3a-b6a6-49f8-be28-3739617ebca6

+++ controller/src/main/java/org/openhds/controller/service/DemRatesService.java
@@ -24,6 +24,6 @@ public interface DemRatesService {
 	List<InMigration> getInMigrationsBetweenInterval(Calendar startDate, Calendar endDate);
 	
 	@Authorized({PrivilegeConstants.VIEW_ENTITY})
-	long daysBetween(Calendar startDate, Calendar endDate); 
+	int daysBetween(Calendar startDate, Calendar endDate); 
 
 }
