+++ controller/src/main/java/org/openhds/controller/service/impl/DemRatesServiceImpl.java
@@ -85,9 +85,9 @@ public class DemRatesServiceImpl implements DemRatesService {
 		return midPoint;
 	}
 	
-	public long daysBetween(Calendar startDate, Calendar endDate) {  
+	public int daysBetween(Calendar startDate, Calendar endDate) {  
 		Calendar date = (Calendar) startDate.clone();  
-		long daysBetween = 0;  
+		int daysBetween = 0;  
 		while (date.before(endDate)) {  
 			date.add(Calendar.DAY_OF_MONTH, 1);  
 		    daysBetween++;
